# contracts/openapi/v1/openapi.yaml
openapi: 3.1.2
...
info:
  title: MASI-NIDS Go Control Core REST API
  version: 1.1.0
  description: |
    Same-origin HTTPS REST API for MOD-CTRL-001. Every list is scope-filtered and
    bounded to 200 records, with an authenticated server cursor. Every business
    mutation requires the OIDC session and CSRF scheme together, an exact
    idempotency_key, Origin/Sec-Fetch-Site validation, and server-side scope
    authorization. A 202 response means a durable operation was accepted or
    prepared; it never means an external side effect succeeded.
  license: {name: Apache-2.0, identifier: Apache-2.0}
servers:
  - {url: https://control-core.local, description: Same-origin production endpoint}
security:
  - oidc: []
paths:
...
  /api/effects/decisions:
    get:
      operationId: listDecisions
      parameters: [{$ref: '#/components/parameters/Cursor'}, {$ref: '#/components/parameters/PageSize'}]
      responses: {'200': {$ref: '#/components/responses/Projection200'}}
    post:
      operationId: recordDecision
      security: [{oidc: [], csrf: []}]
      requestBody: {$ref: '#/components/requestBodies/RecordDecision'}
      responses:
        '201': {description: terminal authorization decision recorded, content: {application/json: {schema: {$ref: '#/components/schemas/Decision'}}}}
        '403': {$ref: '#/components/responses/Error403'}
        '409': {$ref: '#/components/responses/Error409'}
...
components:
  securitySchemes:
    oidc:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://idp.local/authorize
          tokenUrl: https://idp.local/token
          scopes:
            openid: OpenID Connect identity
            masi:operator: scoped operator
            masi:platform-admin: scoped platform administrator
            masi:analyst: scoped analyst
            masi:auditor: scoped auditor
            plugin.statistics.run: on-demand statistics execution
    csrf: {type: apiKey, in: header, name: X-CSRF-Token}
  parameters:
    Cursor: {name: cursor, in: query, required: false, schema: {type: string, maxLength: 512, pattern: '^[A-Za-z0-9_-]+\.[A-Za-z0-9_-]+$'}}
    PageSize: {name: page_size, in: query, required: false, schema: {type: integer, minimum: 1, maximum: 200, default: 50}}
...
  requestBodies:
...
    RecordDecision:
      required: true
      content: {application/json: {schema: {$ref: '#/components/schemas/RecordDecisionRequest'}}}
...
  responses:
    Projection200: {description: bounded scope-filtered projection, content: {application/json: {schema: {$ref: '#/components/schemas/Projection'}}}}
...
    Error403: {description: authorization denied, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
...
    Error409: {description: CAS, idempotency, immutable identity, or stale fence conflict, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
...
  schemas:
    Identity: {type: string, minLength: 1, maxLength: 128, pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]*$'}
...
    ReasonCode: {type: string, pattern: '^[A-Z][A-Z0-9_]{0,63}$'}
    IdempotencyKey: {type: string, minLength: 1, maxLength: 128, pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]*$'}
    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties: {error: {$ref: '#/components/schemas/ReasonCode'}}
...
    Projection:
      type: object
      additionalProperties: false
      required: [schema_version, projection_type, resource_kind, cursor, page_size, total_count, items, generation, session_scope, authorized_scope, sse_event, actor_ref, reason_code, trace_id]
      properties:
        schema_version: {const: masi-web-projection/v1}
        projection_type: {enum: [current, desired, observed, stale, hold, unknown, reconciling]}
        resource_kind: {$ref: '#/components/schemas/Identity'}
        resource_id: {$ref: '#/components/schemas/Identity'}
        cursor: {type: string, maxLength: 512}
        page_size: {type: integer, minimum: 1, maximum: 200}
        total_count: {type: integer, minimum: 0}
        items:
          type: array
          maxItems: 200
          items: {type: object, maxProperties: 64}
        generation: {type: integer, minimum: 1}
        session_scope: {type: string, minLength: 1, maxLength: 4096}
        authorized_scope: {type: string, minLength: 1, maxLength: 4096}
        sse_event: {type: [object, 'null']}
        actor_ref: {$ref: '#/components/schemas/Identity'}
        reason_code: {$ref: '#/components/schemas/ReasonCode'}
        trace_id: {type: string, maxLength: 128}
...
    RecordDecisionRequest:
      type: object
      additionalProperties: false
      required: [proposal_id, decision, reason_code, idempotency_key]
      properties:
        proposal_id: {$ref: '#/components/schemas/Identity'}
        decision: {enum: [approve, reject]}
        reason_code: {$ref: '#/components/schemas/ReasonCode'}
        idempotency_key: {$ref: '#/components/schemas/IdempotencyKey'}
...
    Decision: {type: object, minProperties: 10, maxProperties: 20}
...